django_portfolio
This repository contains the source code for my personal portfolio website, built using HTML, CSS, JavaScript, and the Django web framework. The site features a responsive design and includes sections for showcasing my projects, skills, and experience. Users can browse my portfolio and contact me directly through a contact form. I built this website as a showcase of my skills and to provide an online presence for myself as a developer. Feel free to browse the code, contribute, or fork the repository to create your own portfolio website!